fix: converted css to tailwind
