--- packages/components/src/components/carousel-item/carousel-item.component.ts
@@ -1,38 +1,0 @@
-import { html } from 'lit';
-import SolidElement from '../../internal/solid-element.js';
-import styles from './carousel-item.styles.js';
-import type { CSSResultGroup } from 'lit';
-
-/**
- * @summary A carousel item represent a slide within a [carousel](/components/carousel).
- *
- * @since 2.0
- * @status experimental
- *
- * @slot - The carousel item's content..
- *
- * @cssproperty --aspect-ratio - The slide's aspect ratio. Inherited from the carousel by default.
- *
- */
-export default class SdCarouselItem extends SolidElement {
-  static styles: CSSResultGroup = styles;
-
-  static isCarouselItem(node: Node) {
-    return node instanceof Element && node.getAttribute('aria-roledescription') === 'slide';
-  }
-
-  connectedCallback() {
-    super.connectedCallback();
-    this.setAttribute('role', 'group');
-  }
-
-  render() {
-    return html` <slot></slot> `;
-  }
-}
-
-declare global {
-  interface HTMLElementTagNameMap {
-    'sd-carousel-item': SdCarouselItem;
-  }
-}
