+++ packages/components/src/components/carousel-item/carousel-item.stories.ts
@@ -1,7 +1,5 @@
 import '../../solid-components';
 import { storybookDefaults, storybookTemplate } from '../../../scripts/storybook/helper';
-import { withActions } from '@storybook/addon-actions/decorator';
-
 const { argTypes, args, parameters } = storybookDefaults('sd-carousel-item');
 const { generateTemplate } = storybookTemplate('sd-carousel-item');
 
@@ -10,8 +8,7 @@ export default {
   component: 'sd-carousel-item',
   args,
   argTypes,
-  parameters: { ...parameters },
-  decorators: [withActions] as any
+  parameters: { ...parameters }
 };
 
 /**
