+++ packages/components/src/components/carousel-item/carousel-item.ts
@@ -1,4 +1,40 @@
-import SdCarouselItem from './carousel-item.component.ts';
-export * from './carousel-item.component.ts';
-export default SdCarouselItem;
-SdCarouselItem.define('sd-carousel-item');
+import { customElement } from 'lit/decorators.js';
+import { html } from 'lit';
+import SolidElement from '../../internal/solid-element.js';
+import styles from './carousel-item.styles.js';
+import type { CSSResultGroup } from 'lit';
+
+/**
+ * @summary A carousel item represent a slide within a [carousel](/components/carousel).
+ *
+ * @since 2.0
+ * @status experimental
+ *
+ * @slot - The carousel item's content..
+ *
+ * @cssproperty --aspect-ratio - The slide's aspect ratio. Inherited from the carousel by default.
+ *
+ */
+@customElement('sd-carousel-item')
+export default class SdCarouselItem extends SolidElement {
+  static styles: CSSResultGroup = styles;
+
+  static isCarouselItem(node: Node) {
+    return node instanceof Element && node.getAttribute('aria-roledescription') === 'slide';
+  }
+
+  connectedCallback() {
+    super.connectedCallback();
+    this.setAttribute('role', 'group');
+  }
+
+  render() {
+    return html` <slot></slot> `;
+  }
+}
+
+declare global {
+  interface HTMLElementTagNameMap {
+    'sd-carousel-item': SdCarouselItem;
+  }
+}
