+++ packages/components/src/components/carousel/carousel.styles.ts
